+++ examples/copyContentToProjectFolder/applications/example/app/examples/tools/forkWorker.json/worker.js
@@ -1,5 +1,5 @@
 module.exports = function(client, callback) {
   console.log('Message from forked worker');
-  console.dir(client);
+  console.dir(Object.keys(client));
   callback();
 }